@@ -3,42 +3,41 @@
 // </copyright>
 
 using IdleOfTheAgesLib.DependencyInjection;
-
 using System.Reflection;
 
-namespace IdleOfTheAgesLib {
+namespace IdleOfTheAgesLib;
+
+/// <summary>
+/// Contains all relevant mod data.
+/// </summary>
+public interface IModObject {
     /// <summary>
-    /// Contains all relevant mod data.
+    /// Gets a logger to log data to.
     /// </summary>
-    public interface IModObject {
-        /// <summary>
-        /// Gets a logger to log data to.
-        /// </summary>
-        ILogger Logger { get; }
+    ILogger Logger { get; }
 
-        /// <summary>
-        /// Gets the instance of the mod's initializer class.
-        /// </summary>
-        IMod Mod { get; }
+    /// <summary>
+    /// Gets the instance of the mod's initializer class.
+    /// </summary>
+    IMod Mod { get; }
 
-        /// <summary>
-        /// Gets the namespace of the mod.
-        /// </summary>
-        string Namespace { get; }
+    /// <summary>
+    /// Gets the namespace of the mod.
+    /// </summary>
+    string Namespace { get; }
 
-        /// <summary>
-        /// Gets the service library containing this mod's services.
-        /// </summary>
-        IServiceLibrary ServiceLibrary { get; }
+    /// <summary>
+    /// Gets the mod's private service library.
+    /// </summary>
+    IServiceLibrary ServiceLibrary { get; }
 
-        /// <summary>
-        /// Gets the service registry from this mod to register services to.
-        /// </summary>
-        IServiceRegistry ServiceRegistry { get; }
+    /// <summary>
+    /// Gets the service registry from this mod to register services to.
+    /// </summary>
+    IServiceRegistry ServiceRegistry { get; }
 
-        /// <summary>
-        /// Gets the mod's assembly.
-        /// </summary>
-        Assembly ModAssembly { get; }
-    }
+    /// <summary>
+    /// Gets the mod's assembly.
+    /// </summary>
+    Assembly ModAssembly { get; }
 }
