@@ -2,10 +2,10 @@
 // Copyright (c) DaanStout. All rights reserved.
 // </copyright>
 
-namespace IdleOfTheAgesLib.Inventory {
-    /// <summary>
-    /// Inherit from this class to add a bank tab.
-    /// </summary>
-    public interface IStorageGroup {
-    }
+namespace IdleOfTheAgesLib.Inventory;
+
+/// <summary>
+/// Inherit from this class to add a bank tab.
+/// </summary>
+public interface IStorageGroup {
 }
