@@ -4,30 +4,16 @@
 
 using IdleOfTheAgesLib.Models;
 
-namespace IdleOfTheAgesLib {
+namespace IdleOfTheAgesLib;
+
+/// <summary>
+/// Allows for loading in <see cref="ModData"/> objects.
+/// </summary>
+public interface IDataLoader {
     /// <summary>
-    /// Allows for loading in <see cref="ModData"/> objects.
+    /// Loads in the files in the mod's 'data' folder.
     /// </summary>
-    public interface IDataLoader {
-        /// <summary>
-        /// Loads in a <see cref="ModData"/> object from the mod folder.
-        /// </summary>
-        /// <param name="pathSegments">The path to the json file.</param>
-        /// <returns>A <see cref="Result"/> object to check if the call was successful.</returns>
-        Result LoadData(params string[] pathSegments);
-
-        /// <summary>
-        /// Registers textures from this mod to the game.
-        /// </summary>
-        /// <param name="pathSegments">The root path.</param>
-        /// <returns>A <see cref="Result"/> object to check if the call was successful.</returns>
-        Result RegisterTextures(params string[] pathSegments);
-
-        /// <summary>
-        /// Loads the language files in the given path.
-        /// </summary>
-        /// <param name="pathSegments">The path the language files are located at.</param>
-        /// <returns>A <see cref="Result"/> object to check if the call was successful.</returns>
-        Result LoadLanguages(params string[] pathSegments);
-    }
+    /// <param name="mod">The mod to load the data in for..</param>
+    /// <returns><see langword="true"/> if the data was successfully loaded..</returns>
+    Result LoadData(IModObject mod);
 }
