@@ -2,12 +2,11 @@
 // Copyright (c) DaanStout. All rights reserved.
 // </copyright>
 
-namespace IdleOfTheAgesLib.User
-{
-    /// <summary>
-    /// A service that holds all relevant information about the player.
-    /// <para>For information about the player, see <see cref="IUserService"/>.</para>
-    /// </summary>
-    public interface IPlayerService {
-    }
+namespace IdleOfTheAgesLib.User;
+
+/// <summary>
+/// A service that holds all relevant information about the player.
+/// <para>For information about the user, see <see cref="IUserService"/>.</para>
+/// </summary>
+public interface IPlayerService {
 }
