@@ -2,11 +2,11 @@
 // Copyright (c) DaanStout. All rights reserved.
 // </copyright>
 
-namespace IdleOfTheAgesLib.User {
-    /// <summary>
-    /// A service that holds all relevant information about the user.
-    /// <para>For information about the player, see <see cref="IPlayerService"/>.</para>
-    /// </summary>
-    public interface IUserService {
-    }
+namespace IdleOfTheAgesLib.User;
+
+/// <summary>
+/// A service that holds all relevant information about the user.
+/// <para>For information about the player, see <see cref="IPlayerService"/>.</para>
+/// </summary>
+public interface IUserService {
 }
