@@ -4,142 +4,134 @@
 
 using IdleOfTheAgesLib;
 using IdleOfTheAgesLib.Models;
-using IdleOfTheAgesLib.Models.ModJsonData;
+using IdleOfTheAgesLib.Models.JsonConverters;
 using IdleOfTheAgesLib.Skills;
 using IdleOfTheAgesLib.Translation;
-using IdleOfTheAgesLib.UI;
-
+using IdleOfTheAgesLib.UI.Parsing;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace IdleOfTheAgesRuntime {
+namespace IdleOfTheAgesRuntime;
+
+/// <summary>
+/// Allows for loading in data into the game.
+/// </summary>
+public class DataLoader : IDataLoader {
+    private static readonly JsonSerializerSettings SETTINGS = new() {
+        ContractResolver = new DefaultContractResolver {
+            NamingStrategy = new SnakeCaseNamingStrategy(),
+        },
+        Converters = [new ModDataConverter()],
+    };
+
     /// <summary>
-    /// Allows for loading in data into the game.
+    /// Extensions we explicitly ignore during the loading of data and don't log errors for.
     /// </summary>
-    public class DataLoader : IDataLoader {
-        private static readonly JsonSerializerSettings Settings = new JsonSerializerSettings() {
-            ContractResolver = new DefaultContractResolver {
-                NamingStrategy = new SnakeCaseNamingStrategy(),
-            },
-        };
+    private static readonly string[] EXPLICITLY_IGNORED_EXTENSIONS = [".import"];
 
-        private readonly IModObject modObject;
-        private readonly IAgeService ageService;
-        private readonly ISkillService skillService;
-        private readonly ITextureLibrary textureLibrary;
-        private readonly ITranslationService translationService;
-        private readonly IPageService pageGroupService;
+    private readonly ISkillService skillService;
+    private readonly ITranslationService translationService;
+    private readonly IActionLibrary actionLibrary;
+    private readonly IParserService parserService;
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="DataLoader"/> class.
-        /// </summary>
-        /// <param name="modObject">The mod the data is loaded form.</param>
-        /// <param name="ageService">The age service to load ages into.</param>
-        /// <param name="skillService">The skill service to load skills into.</param>
-        /// <param name="textureLibrary">The texture library to load textures into.</param>
-        /// <param name="translationService">The translation service to load translation files into.</param>
-        /// <param name="pageGroupService">THe page group service to load page info into.</param>
-        public DataLoader(IModObject modObject, IAgeService ageService, ISkillService skillService, ITextureLibrary textureLibrary, ITranslationService translationService, IPageService pageGroupService) {
-            this.modObject = modObject;
-            this.ageService = ageService;
-            this.skillService = skillService;
-            this.textureLibrary = textureLibrary;
-            this.translationService = translationService;
-            this.pageGroupService = pageGroupService;
+    /// <summary>
+    /// Initializes a new instance of the <see cref="DataLoader"/> class.
+    /// </summary>
+    /// <param name="skillService">The skill service to load skills into.</param>
+    /// <param name="translationService">The translation service to load translation files into.</param>
+    /// <param name="actionLibrary">The action library to load actions into.</param>
+    /// <param name="parserService">The parser service to load UI files into.</param>
+    public DataLoader(ISkillService skillService, ITranslationService translationService, IActionLibrary actionLibrary, IParserService parserService) {
+        this.skillService = skillService;
+        this.translationService = translationService;
+        this.actionLibrary = actionLibrary;
+        this.parserService = parserService;
+    }
+
+    /// <inheritdoc/>
+    public Result LoadData(IModObject mod) {
+        ResultBuilder builder = new();
+
+        var path = Path.Combine(Directory.GetCurrentDirectory(), "Resources", "Mods", mod.Namespace, "Assets");
+
+        if (!Directory.Exists(path))
+            return (false, "Directory does not exist!", new ArgumentException($"No data directory for mod {mod.Namespace}"));
+
+        LoadDataInDirectory(mod.Namespace, path, builder);
+
+        return builder.Build(true);
+    }
+
+    private void LoadDataInDirectory(string modNamespace, string directory, ResultBuilder builder) {
+        foreach (var file in Directory.GetFiles(directory)) {
+            LoadFile(modNamespace, file, builder);
         }
 
-        /// <inheritdoc/>
-        public Result LoadData(params string[] pathSegments) {
-            var path = Path.Combine(UnityEngine.Application.streamingAssetsPath, "Mods", modObject.Namespace, Path.Combine(pathSegments));
+        foreach (var subDirectory in Directory.GetDirectories(directory)) {
+            LoadDataInDirectory(modNamespace, subDirectory, builder);
+        }
+    }
 
-            if (!File.Exists(path)) {
-                return (false, $"File does not exist: {path}");
-            }
+    private void LoadFile(string modNamespace, string file, ResultBuilder builder) {
+        var extension = Path.GetExtension(file);
 
-            var json = File.ReadAllText(path);
-
-            ModData data;
-            try {
-                data = JsonConvert.DeserializeObject<ModData>(json, Settings) ?? throw new FormatException("Cannot deserialize json into mod data file!");
-            } catch (Exception e) {
-                return (false, $"Error parsing the json for file at path: {path}", e);
-            }
-
-            ProcessData(data, data.PageGroups, pageGroup => pageGroupService.RegisterPageGroup(pageGroup));
-            ProcessData(data, data.Pages, page => pageGroupService.RegisterPage(page));
-            ProcessData(data, data.Ages, age => ageService.RegisterAge(age));
-            ProcessData(data, data.Skills, skill => skillService.RegisterSkillData(skill));
-
-            return true;
+        if (EXPLICITLY_IGNORED_EXTENSIONS.Contains(extension)) {
+            return;
         }
 
-        /// <inheritdoc/>
-        public Result LoadLanguages(params string[] pathSegments) {
-            var path = Path.Combine(UnityEngine.Application.streamingAssetsPath, "Mods", modObject.Namespace, Path.Combine(pathSegments));
+        switch (extension) {
+            case ".txt":
+                var language = Path.GetFileNameWithoutExtension(file);
 
-            if (!Directory.Exists(path)) {
-                return (false, "Directory does not exist");
-            }
-
-            List<string> failedFiles = new List<string>();
-
-            foreach (var file in Directory.GetFiles(path)) {
-                if (Enum.TryParse<Language>(Path.GetFileNameWithoutExtension(file), true, out var language)) {
+                if (Languages.IsLanguageSupported(language)) {
                     translationService.LoadLanguagePath(language, file);
                 } else {
-                    failedFiles.Add(Path.GetFileName(path));
-                }
-            }
-
-            if (failedFiles.Count > 0) {
-                return (false, $"Failed loading {failedFiles.Count} files!{failedFiles.Aggregate(string.Empty, (acc, str) => Environment.NewLine + acc + str)}");
-            } else {
-                return true;
-            }
-        }
-
-        /// <inheritdoc/>
-        public Result RegisterTextures(params string[] pathSegments) {
-            var rootPath = Path.Combine(UnityEngine.Application.streamingAssetsPath, "Mods", modObject.Namespace, Path.Combine(pathSegments));
-
-            if (!Directory.Exists(rootPath)) {
-                return (false, "Directory does not exist");
-            }
-
-            Queue<string> directories = new Queue<string>();
-            directories.Enqueue(rootPath);
-
-            while (directories.Count > 0) {
-                var directory = directories.Dequeue();
-
-                foreach (var subFolder in Directory.GetDirectories(directory)) {
-                    directories.Enqueue(subFolder);
+                    builder.AddError($"Failed loading language file: {file}. The language is not supported.");
                 }
 
-                foreach (var filePath in Directory.GetFiles(directory)) {
-                    var fileExtension = Path.GetExtension(filePath);
+                break;
+            case ".json":
+                var rawJson = File.ReadAllText(file);
 
-                    if (fileExtension != ".png" && fileExtension != ".jpg") {
-                        continue;
+                try {
+                    var data = JsonConvert.DeserializeObject<ModData>(rawJson, SETTINGS);
+
+                    if (data == null) {
+                        builder.AddError($"Error while parsing json at file path: {file}");
+                        break;
                     }
 
-                    textureLibrary.RegisterTextures($"{modObject.Namespace}:{Path.GetFileNameWithoutExtension(filePath)}", filePath);
+                    LoadModData(data, builder);
+                } catch (Exception e) {
+                    builder.AddError(($"Error while parsing json at file path: {file}", e));
                 }
-            }
 
-            return true;
+                break;
+            case ".html":
+                var fileName = Path.GetFileNameWithoutExtension(file);
+                fileName = $"{modNamespace}:{fileName}";
+
+                builder.AddResult(parserService.RegisterFile(fileName, file));
+
+                break;
+            case ".png":
+                break;
+            default:
+                builder.AddError($"File extensions '{extension}' is not supported");
+                break;
+        }
+    }
+
+    private void LoadModData(ModData data, ResultBuilder builder) {
+        foreach (var skill in data.Skills) {
+            builder.AddResult(skillService.RegisterSkillData(skill));
         }
 
-        private static void ProcessData<T>(ModData data, IEnumerable<T> elements, Action<T> processFunc) where T : BaseDataElement {
-            foreach (var item in elements) {
-                item.Namespace = data.Namespace;
-                processFunc(item);
-            }
+        foreach (var action in data.Actions) {
+            builder.AddResult(actionLibrary.RegisterAction(action));
         }
     }
 }
