@@ -4,11 +4,11 @@
 
 using IdleOfTheAgesLib.User;
 
-namespace IdleOfTheAgesRuntime.User {
-    /// <summary>
-    /// A service that holds all relevant information about the user.
-    /// <para>For information about the player, see <see cref="IPlayerService"/>.</para>
-    /// </summary>
-    public class UserService : IUserService {
-    }
+namespace IdleOfTheAgesRuntime.User;
+
+/// <summary>
+/// A service that holds all relevant information about the user.
+/// <para>For information about the player, see <see cref="IPlayerService"/>.</para>
+/// </summary>
+public class UserService : IUserService {
 }
