@@ -13,4 +13,4 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Is part of base classes to prevent duplication.", Scope = "type", Target = "~T:IdleOfTheAgesLib.Models.ModJsonData.ThumbnailDataElement")]
 [assembly: SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Is part of base classes to prevent duplication.", Scope = "type", Target = "~T:IdleOfTheAgesLib.Models.ModJsonData.VisisbleDataElement")]
 [assembly: SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Is a generic base class for use.", Scope = "type", Target = "~T:IdleOfTheAgesLib.Result`1")]
-[assembly: SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Is a generic base class that is meant to be inherited from.", Scope = "type", Target = "~T:IdleOfTheAgesLib.UI.Elements.Element`2")]
+[assembly: SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Is a generic base class for use.", Scope = "type", Target = "~T:IdleOfTheAgesLib.UI.Elements.ElementManager`1")]
